IDE-4021 reconstruct hook project and hook deploy project tests
