+++ ui-tests/base/com.liferay.ide.ui.hook.deploy.base/.gitignore
