+++ ui-tests/com.liferay.ide.ui.testing-repository/category.xml
@@ -11,4 +11,5 @@
 	<bundle id="com.liferay.ide.ui.server.wizard.base" version="0.0.0" />
 	<bundle id="com.liferay.ide.ui.layouttpl.deploy.base" version="0.0.0" />
 	<bundle id="com.liferay.ide.ui.theme.deploy.base" version="0.0.0" />
+	<bundle id="com.liferay.ide.ui.hook.deploy.base" version="0.0.0" />
 </site>
