+++ ui-tests/tools/com.liferay.ide.ui.hook/com.liferay.ide.ui.hook.wizard/com.liferay.ide.ui.hook.wizard.create/build.properties
@@ -1,0 +1,4 @@
+source.. = src/
+output.. = target/classes/
+bin.includes = META-INF/,\
+               .
